% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/censor_at.r
\name{censor_at}
\alias{censor_at}
\title{Censor time to event end point at a certain time.}
\usage{
censor_at(time = NULL, status = NULL, censoring_time = NULL)
}
\arguments{
\item{time}{time variable}

\item{status}{status variable}

\item{censoring_time}{censoring time}
}
\value{
A data frame to be used with \code{\link{cbind}}
}
\description{
Censor time to event end point at a certain time. Implemented only for
"right" type of censored data.
}
\examples{
time   <- c(100,150)
status <- c(  0,  1)
(original <- data.frame(time, status))
censor_at(time = time, status = status, censoring_time = 160)
censor_at(time = time, status = status, censoring_time = 150)
censor_at(time = time, status = status, censoring_time = 125)
censor_at(time = time, status = status, censoring_time = 75)

time   <- c(100,150)
status <- c(  1,  0)
(original <- data.frame(time, status))
censor_at(time = time, status = status, censoring_time = 160)
censor_at(time = time, status = status, censoring_time = 150)
censor_at(time = time, status = status, censoring_time = 125)
censor_at(time = time, status = status, censoring_time = 75)
}

