% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprisk.R
\name{comprisk}
\alias{comprisk}
\title{Plot and compare cumulative/crude incidence in a competing risk
setting}
\usage{
comprisk(
  time = NULL,
  status = NULL,
  strata = NULL,
  time_unit = c("days", "weeks", "months", "years"),
  time_by = NULL,
  plot = TRUE,
  plot_ci = FALSE,
  ylab = NULL,
  xlab = NULL,
  ylim = c(0L, 1L),
  xlim = NULL,
  test = c("gray", "none"),
  col_status = seq_len(nlevels(status) - 1),
  col_strata = seq_len(nlevels(strata)),
  cex_test = par("cex") * 0.8,
  cex_legend = par("cex") * 0.8,
  plot_grid = TRUE
)
}
\arguments{
\item{time}{survival time variable}

\item{status}{integer with 0 = censored or factor with first level
= censored}

\item{strata}{Stratifying variable (optional)}

\item{time_unit}{Time unit of x axis}

\item{time_by}{Time step x axis (in days) @param quantile_probs}

\item{plot}{plot (default = TRUE) or only return estimates?}

\item{plot_ci}{plot confidence intervalse for estimates (for
univariate unstacked and multivariate plots)}

\item{ylab}{Y-axis label.}

\item{xlab}{X-axis label. If NULL a suitable default based on
time_unit will be provided}

\item{ylim}{Y-axis limit. Default to c(0,1) be provided @param}

\item{xlim}{X-axis limit. If NULL a suitable default based on
time_unit will be provided}

\item{test}{tests: 'none' = don't plot tests, 'gray' = gray test}

\item{col_status}{color for exit status (mainly for univariate
plotting)}

\item{col_strata}{color for group status (mainly for bivariate
plotting)}

\item{cex_test}{cex parameter for test string}

\item{cex_legend}{cex parameter for legend string}

\item{plot_grid}{plot grid @param plot_n_at_risk Logical value:}
}
\description{
Plot and compare cumulative/crude incidence in a competing risk
setting
}
\examples{
\dontrun{

library(lbdatasets)
(ci.fns <- with(mros, cmprsk::cuminc(time, status, cutbmd))$Tests)

## Full example
par(mfrow = c(1,2))
res <- comprisk(time = mros$time,
                status = factor(mros$status, levels = 0:2,
                                labels = c("Censored", "Fractured", "Death")),
                strata = factor(mros$cutbmd, levels = 1:2, labels = c('LowBMD', 'HighBMD')),
                time_by = 50, ylim = c(0, 0.25), col = c('black', 'red'))
}

}
