% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contal.R
\name{contal}
\alias{contal}
\title{Apply contal o quigley method}
\usage{
contal(time = NULL, status = NULL, marker = NULL)
}
\arguments{
\item{time}{time variable}

\item{status}{status variable}

\item{marker}{marker variable}
}
\description{
Apply contal o quigley method
}
\examples{

## Klein Moeschberg example pag 274
data(kidtran, package = "KMsurv")
kidtran <- within(kidtran, {
    gender <- factor(gender, levels = c(1,2), labels = c('male', 'female'))
    race <- factor(race, levels = c(1,2), labels = c('white', 'black'))
})
k_spl <- split(kidtran, f = list(kidtran$race, kidtran$gender))
k_spl <- k_spl[c(2,1,4,3)]
expanded <- lapply(k_spl, function(x){
    contal(time = x$time,
           status = x$delta,
           marker = x$age)
})
expanded[1] 

## expanded_survmisc <- lapply(k_spl, function(x){
##     df = data.frame(time = x$time,
##                     status = x$delta,
##                     marker = x$ag)
##     cox <- survival::coxph(Surv(time = time, event = status) ~ marker,
##                            data = df)
##     survMisc::cutp(cox)
## })

compact <- do.call(rbind, lapply(expanded, function(x) x$optimal))
compact

}
