% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fup_completeness.R
\name{fup_completeness}
\alias{fup_completeness}
\title{Calculate C index of of follow up completeness}
\usage{
fup_completeness(time = NULL, status = NULL, cutoff = seq(1, max(time),
  length = 10), strata = NULL)
}
\arguments{
\item{time}{Follow up (in days?)}

\item{status}{event indicator}

\item{cutoff}{(in days?)}

\item{strata}{group}
}
\value{
A data.frame with a global C and a C for each group
}
\description{
Calculate C index of of follow up completeness.
}
\examples{

time   <- c(180, 12, 240, 250 )
status <- c(  0,  1,   0, 1   )
group  <- c("A","A", "B", "B" )

## example: 
## quantify fup completeness to 200 days (eg minimum potential
## follow up in a hypotethic prospective trial)

fup_completeness(time = time, status = status,
                 cutoff = seq(150, 250, 10),
                 strata = group)

}
\references{
Clark T., Altman D., De Stavola B. (2002),
Quantification of the completeness of follow-up. Lancet 2002;
359: 1309-10
}
