% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ill_death}
\alias{ill_death}
\title{illness death multi-state model plot}
\usage{
ill_death(
  entry_lab = "Entry",
  illness_lab = "Relapse",
  death_lab = "Death",
  recover = FALSE,
  box = FALSE
)
}
\arguments{
\item{entry_lab}{char entry state label}

\item{illness_lab}{char illness state label}

\item{death_lab}{char death state label}

\item{recover}{draw transition between illness and entry states}

\item{box}{add graphics::box}
}
\description{
illness death multi-state model plot
}
\examples{
ill_death(recover = TRUE, box = TRUE)
}
