% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km.R
\name{km_legend}
\alias{km_legend}
\title{Make a suitable legend command for km}
\usage{
km_legend(title, levels, colors, lty, lwd)
}
\arguments{
\item{title}{legend title}

\item{levels}{factor levels}

\item{colors}{colors}

\item{lty}{lty}

\item{lwd}{lwd}
}
\description{
Make a suitable legend command for km
}

