% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fup_stats.R
\name{median_fup}
\alias{median_fup}
\title{utility function around fup_stats}
\usage{
median_fup(
  os_time = NULL,
  os_status = NULL,
  strata = NULL,
  type = c("revkm", "raw", "all"),
  time_unit = c("days", "weeks", "months", "years")
)
}
\arguments{
\item{os_time}{time}

\item{os_status}{event indicator}

\item{strata}{group}

\item{type}{which method: 'revkm' for reverse kaplan meier or 'raw'
(eg look at censored observation time only)}

\item{time_unit}{Time unit table returned}
}
\description{
utility function around fup_stats
}
