% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_test.R
\name{ph_test}
\alias{ph_test}
\title{A wrapper for cox.zph which plot Schoenfeld residuals of variables
violating PH assumption}
\usage{
ph_test(cox_mod, plot_signif = TRUE, mfrow = NULL)
}
\arguments{
\item{cox_mod}{a coxph object}

\item{plot_signif}{plot Schoenfeld residuals for variables violating
PH assumption?}

\item{mfrow}{optimized (optional) mfrow}
}
\description{
A wrapper for cox.zph which plot Schoenfeld residuals of variables
violating PH assumption
}
