% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_at.R
\name{status_at}
\alias{status_at}
\title{Return patients status at given time}
\usage{
status_at(
  start_date = NULL,
  prog_date = NULL,
  death_date = NULL,
  status_time = NULL
)
}
\arguments{
\item{start_date}{time variable}

\item{prog_date}{status variable}

\item{death_date}{death variable}

\item{status_time}{censoring time}
}
\value{
a factor vector
}
\description{
Return patients status at given time. Implemented for
right-censored dataset.
}
\examples{
db <- data.frame(start_date  = as.Date(c("1900-01-01", "1900-01-01", "1900-01-01", "1900-01-01",           NA)),
                 prog_date   = as.Date(c("1900-03-01", "1900-03-01",           NA,           NA,           NA)),
                 death_date  = as.Date(c("1900-06-01",           NA, "1900-06-01",           NA,           NA)))
with(db, status_at(start_date, prog_date, death_date, status_time = 10))
with(db, status_at(start_date, prog_date, death_date, status_time = 100))
with(db, status_at(start_date, prog_date, death_date, status_time = 200))
}
