% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fup_stats.R
\name{fup_stats}
\alias{fup_stats}
\title{Calculate follow up statistics}
\usage{
fup_stats(
  time = NULL,
  status = NULL,
  cutoff = seq(1, max(time), length = 10),
  strata = NULL,
  time_unit = c("months", "days", "weeks", "years"),
  time_by = NULL
)
}
\arguments{
\item{time}{Follow up (in days?)}

\item{status}{event indicator}

\item{cutoff}{(in days?)}

\item{strata}{group}

\item{time_unit}{Time unit table returned}

\item{time_by}{Time step for completeness table axis (in days)}
}
\value{
A data.frame with a global C and a C for each group
}
\description{
Calculate median follow up time (reverse KM method) and C index 
for follow up completeness.
}
\examples{

time   <- c(180,  12, 240, 250 )
status <- c(  0,   1,   0,   1 )
group  <- c("A", "A", "B", "B" )

## example: 
## quantify fup completeness to 200 days (eg minimum potential
## follow up in a hypotethic prospective trial)

fup_stats(time = time, status = status,
          cutoff = seq(150, 250, 10),
          strata = group)

}
\references{
Clark T., Altman D., De Stavola B. (2002),
    Quantification of the completeness of follow-up. Lancet 2002;
    359: 1309-10
}
