% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km.R
\name{km}
\alias{km}
\title{Plots an 'enhanced' Kaplan-Meier plot, with base graphics package.}
\usage{
km(time = NULL, status = NULL, strata = NULL, time_unit = c("days",
  "weeks", "months", "years"), time_by = NULL, main = "", ylab = NULL,
  xlab = NULL, ylim = c(0L, 1L), xlim = NULL, reverse = FALSE,
  conf_int = NULL, test = c("logr", "hr", "both", "none"),
  plot_n_at_risk = TRUE, gr_legend = NULL, ...)
}
\arguments{
\item{time}{survival time variable}

\item{status}{survival indicator variable}

\item{strata}{Stratifying variable (optional)}

\item{time_unit}{Time unit of x axis}

\item{time_by}{Time step x axis (in days)}

\item{main}{Graph main title}

\item{ylab}{Y-axis label.}

\item{xlab}{X-axis label. If NULL a suitable default based on
time_unit will be provided}

\item{ylim}{Y-axis limit. Default to c(0,1) be provided}

\item{xlim}{X-axis limit. If NULL a suitable default based on
time_unit will be provided}

\item{reverse}{plot cumulative events}

\item{conf_int}{logical ... Plot confidence intervall? If NULL
confidence interval are plotted only if strata has two or more
levels}

\item{test}{tests: 'none' = don't plot tests, 'logr' = log-rank test,
'hr'=hazard ratio, 'both' = log-rank test and hazard ratio}

\item{plot_n_at_risk}{Logical value: plot number at risk?}

\item{gr_legend}{Graph command to add legend, as string}

\item{...}{Further \code{\link{lines.survfit}} parameters}
}
\value{
The function plot the graph and return a list with
    Laplan-Meier statistics
}
\description{
This function plots a Kaplan-Meier plot.
}
\details{
The function make the hypothesis that times are measured
    in days; in not leave time_unit not specified
}

