% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km.R
\name{km_legend}
\alias{km_legend}
\title{Make a legend command suitable for km}
\usage{
km_legend(x, y = NULL, levels = NULL, colors = NULL, lty = "solid",
  lwd = 1, title = NULL)
}
\arguments{
\item{x}{same as \code{legend}'s x}

\item{y}{same as \code{legend}'s y}

\item{levels}{factor levels}

\item{colors}{colors}

\item{lty}{lty}

\item{lwd}{lwd}

\item{title}{legend title}
}
\description{
Make a legend command suitable for km
}

