% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtable_goodies.R
\name{xtable.summary.coxph}
\alias{xtable.summary.coxph}
\title{xtable method for summary.coxph}
\usage{
\method{xtable}{summary.coxph}(x, digits = getOption("digits"),
  print.coef = TRUE, print.se.coef = FALSE, print.z = FALSE,
  print.p = TRUE, print.exp.coef = TRUE, print.exp.coef.lci = TRUE,
  print.exp.coef.uci = TRUE, ...)
}
\arguments{
\item{x}{a \code{summary.coxph} object}

\item{digits}{digits}

\item{print.coef}{print beta?}

\item{print.se.coef}{print se(beta)?}

\item{print.z}{print z?}

\item{print.p}{print p?}

\item{print.exp.coef}{print exp(beta)}

\item{print.exp.coef.lci}{print confint(beta) low}

\item{print.exp.coef.uci}{print confint(beta) low}

\item{...}{other things passed to \code{xtable::xtable}}
}
\description{
xtable method for summary.coxph
}

